Movie Picker 2.0!!!
Significant changes to how the program runs now. (Next update will be adding a branch for picking a show over a movie. But I might also just make that a separate script, not sure yet). 

Updates so far: Now you can have a separate column set for movies you've already watched. Also now there is a for loop that can regenerate another suggestion if you don't like the one you got (just to make it less random by adding in human bias haha). Also, if at any point you want to exit the program there is now an input function to do that also in every branch of the program. Excellent! Also since the watched and unwatched columns are now separated, we don't have to search by *No* anymore, since that would be redundant now. This update tripled the code that was in their previously, but I think it was a worthwhile change. Stay tuned for more updates!